refactor(events): move meetup groups into a server-only module
getMeetupGroups had to take the image map as an argument because
events/utils.ts is also imported by client components, which left the same
`getMeetupGroups(locale, (await getMeetupImages()) ?? {})` plumbing repeated
at all three call sites.

Move the whole meetup-group concern into events/meetup-groups.ts, marked
`server-only`, so it can fetch the map itself and pages call
`await getMeetupGroups(locale)`. utils.ts keeps only the pure helpers the
client components actually use (sanitize, TAG_STATUS_MAPPING,
mapEventTranslations), and no longer pulls community-meetups.json into a
module reachable from the browser bundle.

`server-only` also enforces the boundary at build time, which the previous
comment could only describe.

Drops tests/unit/community-meetups.spec.ts.
