+++ app/[locale]/community/events/meetups/page.tsx
@@ -18,12 +18,13 @@ import { getMetadata } from "@/lib/utils/metadata"
 import { getRequiredNamespacesForPage } from "@/lib/utils/translations"
 
 import OrganizerCTA from "../_components/organizer-cta"
-import { getMeetupGroups, mapEventTranslations } from "../utils"
+import { getMeetupGroups } from "../meetup-groups"
+import { mapEventTranslations } from "../utils"
 
 import FilterMeetups from "./_components/filter-meetups"
 import PageJsonLD from "./page-jsonld"
 
-import { getEventsData, getMeetupImages } from "@/lib/data"
+import { getEventsData } from "@/lib/data"
 
 const Page = async (props: { params: Promise<PageParams> }) => {
   const params = await props.params
@@ -43,7 +44,7 @@ const Page = async (props: { params: Promise<PageParams> }) => {
       !e.eventTypes?.includes("conference") &&
       !e.eventTypes?.includes("hackathon")
   )
-  const meetupGroups = getMeetupGroups(locale, (await getMeetupImages()) ?? {})
+  const meetupGroups = await getMeetupGroups(locale)
   // Show API meetups first (sorted by date), then groups (sorted alphabetically)
   const meetups = [...apiMeetups, ...meetupGroups]
