+++ app/[locale]/community/events/page.tsx
@@ -42,10 +42,11 @@ import ContinentTabs from "./_components/continent-tabs"
 import EventCard from "./_components/event-card"
 import FilterEvents from "./_components/filter-events"
 import { SECTION_IDS } from "./constants"
+import { getMeetupGroups } from "./meetup-groups"
 import PageJsonLD from "./page-jsonld"
-import { getMeetupGroups, mapEventTranslations } from "./utils"
+import { mapEventTranslations } from "./utils"
 
-import { getEventsData, getMeetupImages } from "@/lib/data"
+import { getEventsData } from "@/lib/data"
 import ethereumEverywhereLogo from "@/public/images/community/ethereum-everywhere-logo.png"
 import geodeLabsLogo from "@/public/images/community/geode-labs-logo.png"
 import heroImage from "@/public/images/enterprise-eth.png"
@@ -89,7 +90,7 @@ const Page = async (props: { params: Promise<PageParams> }) => {
       !e.eventTypes?.includes("conference") &&
       !e.eventTypes?.includes("hackathon")
   )
-  const meetupGroups = getMeetupGroups(locale, (await getMeetupImages()) ?? {})
+  const meetupGroups = await getMeetupGroups(locale)
   const meetups = [...apiMeetups, ...meetupGroups]
 
   // Continent labels for tabs
